Code cleanup. Resolved a lot of warnings. Removed hyperbolic classes (unusable student code).
git-svn-id: svn://localhost:8080/caleydo_main/trunk@2711 1f7349ae-fd9f-0d40-aeb8-9798e6c0fce3

+++ org.caleydo.core/src/org/caleydo/core/manager/picking/PickingManager.java
@@ -505,7 +505,7 @@ public class PickingManager {
 	}
 
 	private float getDepth(int iZValue) {
-		long depth = (long) iZValue; // large -ve number
+		long depth = iZValue; // large -ve number
 		return (1.0f + ((float) depth / 0x7fffffff));
 		// return as a float between 0 and 1
 	}
